Create STL-to-Aerofoil
Change file path as needed.
This code takes a 3D model of the type 'stl' and converts it into slices presented as 'dat' files. This essentially takes a 3D model of a wing and converts it to aerofoil slices for XFLR5 import.
This is based off of the ideas of YousefAbdulrahman1 who has previously modelled the fuselage effect in XFLR5 using a similar technique.