Unity-Mechanical-Analysis
For a given Blender model and animation, when imported into Unity, this code will calculate the angular velocity and torque of each 'bone' in the animation tree. This does this by taking into account all of the children bones and applying their inertia to the current bone, plus analysing the movement of the current bone. In Unity, the user needs to manually match the bone name and weight.